backend/requirements.txt
fastapi[standard]
torch
pillow
uvicorn[standard]
transformers[torch]
requests
pydantic